Staging file system

+++ src/Services/VideoService.php
@@ -4,6 +4,7 @@ namespace HlsVideos\Services;
 
 use HlsVideos\Models\HlsVideo;
 use FFMpeg;
+use HlsVideos\Models\HlsFolder;
 use Pion\Laravel\ChunkUpload\Handler\HandlerFactory;
 use Pion\Laravel\ChunkUpload\Receiver\FileReceiver;
 use Illuminate\Support\Facades\Storage;
@@ -58,12 +59,22 @@ class VideoService
         return HlsVideo::find($id);
     }
 
-    static function deleteVideo($id)
+    static function deleteVideo($request, $id)
     {
-        return HlsVideo::find($id)->delete();
+        $video = HlsVideo::find($id);
+        $model = $request->modelType ? $request->modelType::find($request->modelId) : null;
+
+        if ($model) {
+            $model->hlsVideos()->detach($id);
+        }
+
+        if (! $video->HlsVideoables()->count() && ! $video->parentFolders()->count()) {
+
+            return HlsVideo::find($id)->delete();
+        }
     }
 
-    public function receiveVideo($request, $model = null)
+    public function receiveVideo($request, $model = null, $folderId = null)
     {
         $receiver = new FileReceiver('file', $request, HandlerFactory::classFromRequest($request));
 
@@ -76,7 +87,7 @@ class VideoService
         if ($fileReceived->isFinished()) { // file uploading is complete / all chunks are uploaded
             $file = $fileReceived->getFile(); // Get file
 
-            $video = $this->handlingUploadedFile($file, $model);
+            $video = $this->handlingUploadedFile($file, $model, folderId: $folderId);
 
             return [
                 "status" => true,
@@ -93,22 +104,7 @@ class VideoService
         ];
     }
 
-    public function receiveFromServer($request, $videoId)
-    {
-        $video = HlsVideo::findOrFail($videoId);
-        $file = $request->file('file');
-        // Store the uploaded file
-        $extension = $extension ?? $file->getClientOriginalExtension();
-        $fileName = "vd.$extension";
-
-        $disk = Storage::disk(config('hls-videos.temp_disk'));
-        $disk->putFileAs((VideoService::getMediaPath()."$videoId"), $file, $fileName);
-
-        $video->qualities()->delete();
-        (new VideoService())->handleVideoQualities($video);
-    }
-
-    public function handlingUploadedFile($file, $model = null, $extension = null, $originalFileName = null, $deleteChunked = true)
+    public function handlingUploadedFile($file, $model = null, $extension = null, $originalFileName = null, $deleteChunked = true, $folderId = null)
     {
 
         // Store the uploaded file
@@ -132,10 +128,36 @@ class VideoService
             'original_file_name' => $originalFileName ?? $file->getClientOriginalName()
         ]);
 
+        $folder = $folderId && $folderId != 'null'
+            ? HlsFolder::find($folderId)
+            : HlsFolder::whereNull('parent_id')->first();
+
+        if ($folder) {
+            $folder->videos()->attach(
+                $video->id,
+                ['title' => $video->original_file_name]
+            );
+        }
+
         if ($model)
             $model->hlsVideos()->attach([$video->id]);
 
         return $video;
+    }
+
+    public function receiveFromServer($request, $videoId)
+    {
+        $video = HlsVideo::findOrFail($videoId);
+        $file = $request->file('file');
+        // Store the uploaded file
+        $extension = $extension ?? $file->getClientOriginalExtension();
+        $fileName = "vd.$extension";
+
+        $disk = Storage::disk(config('hls-videos.temp_disk'));
+        $disk->putFileAs((VideoService::getMediaPath()."$videoId"), $file, $fileName);
+
+        $video->qualities()->delete();
+        (new VideoService())->handleVideoQualities($video);
     }
 
     private function createUniqueVideoUuid(): string
